+++ README.md
@@ -1,5 +1,9 @@
 <p align="center">
-    <img src="https://user-images.githubusercontent.com/1342803/59471117-1c77b300-8e08-11e9-838e-441b280855b3.png" alt="JWTKit">
+    <img 
+        src="https://user-images.githubusercontent.com/1342803/59471117-1c77b300-8e08-11e9-838e-441b280855b3.png" 
+        height="64" 
+        alt="JWTKit"
+    >
     <br>
     <br>
     <a href="https://api.vapor.codes/jwt-kit/master/JWTKit/index.html">
@@ -11,11 +15,11 @@
     <a href="LICENSE">
         <img src="http://img.shields.io/badge/license-MIT-brightgreen.svg" alt="MIT License">
     </a>
-    <a href="https://circleci.com/gh/vapor/jwt-kit">
-        <img src="https://circleci.com/gh/vapor/jwt-kit.svg?style=shield" alt="Continuous Integration">
+    <a href="https://github.com/vapor/jwt/actions">
+        <img src="https://github.com/vapor/jwt/workflows/test/badge.svg" alt="Continuous Integration">
     </a>
     <a href="https://swift.org">
-        <img src="http://img.shields.io/badge/swift-5.0-brightgreen.svg" alt="Swift 5.0">
+        <img src="http://img.shields.io/badge/swift-5.2-brightgreen.svg" alt="Swift 5.2">
     </a>
 </p>
