+++ Sources/JWT/JWTError+Vapor.swift
@@ -1,0 +1,7 @@
+import Vapor
+
+extension JWTError: AbortError {
+    public var status: HTTPResponseStatus {
+        .unauthorized
+    }
+}
